washpr/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.3
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
flower==2.0.1
humanize==4.12.1
idna==3.10
kombu==5.4.2
prometheus_client==0.21.1
prompt_toolkit==3.0.50
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.1
tornado==6.4.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
psycopg2-binary>=2.9
boto3>=1.26.0
django-storages>=1.13.2
gunicorn